:sparkles: feat: added qr to pay transfer API

+++ src/main/java/com/opom/bankingapp/repository/UserRepository.java
@@ -29,5 +29,5 @@ public interface UserRepository {
     Optional<String> findHashedPinById(Long userId);
     void updatePassword(Long userId, String newHashedPassword);
     void updateSelectedAccount(Long userId, int accountId);
-    Optional<Integer> findSelectedAccountIdByUserId(Long userId);
+    Optional<Long> findSelectedAccountIdByUserId(Long userId);
 }
